Fix base64 error when input is multiple of 4 bytes
The base64Decode function was incorrectly re-adding padding when the
input is a multiple of 4 bytes. These input lengths should have no
padding, but it was adding 1, which led to an error:

> illegal base64 data at input byte 4

The base64 package has options to omit padding using
`WithPadding(base64.NoPadding)`[1], and also defines Raw versions of the
pre-defined encoders[2]. This therefore updates the functions to use
these encoders instead, which means they're now just simple wrappers.

[1]https://pkg.go.dev/encoding/base64#Encoding.WithPadding
[2]https://pkg.go.dev/encoding/base64#pkg-variables