Implement URLSafeSerializer
This implements the serialization function to match the Python
URLSafeSerializer. This handles JSON encoding the payload as well as
optionally zlib-compressing it.

The python implementation uses method names dumps and loads, which match
the Python json library. I've therefore chosen to name the methods here
Marshal and Unmarshal to match the Go json library.