Make timestamped signatures compatible with Python.
The Python implementation initially used an epoch of 2011/01/01, but
this was dropped in favour of the standard unix epoch prior to version
1.0.0[1]. This implementation is still using the original epoch, which
means it's not compatible with Python. This therefore updates this
implementation to use the unix epoch to match the Python implementation.

In order to test this change, I've introduced a NowFunc function to
allow overriding the function it uses to obtain the local time, so that
the tests can mock the current time for repeatability. I considered that
this would be also be useful for consumers of this library, so I've made
it an exported value.

[1]https://github.com/pallets/itsdangerous/issues/46