Fix handling for timestamps > 4 bytes.
The Python implementation serialises the timestamp as an 8-byte integer
(using pack with '>Q'[1]), and then trims leading zeroes[2].

This implementation was assuming a fixed 4-byte representation, which
only works because all current timestamps are 4 bytes once leading
zeroes are removed. This updates it to match the Python implementation,
which means it won't break when the current timestamp overflows 4 bytes.

[1]https://github.com/pallets/itsdangerous/blob/2.2.0/src/itsdangerous/encoding.py#L44
[2]https://github.com/pallets/itsdangerous/blob/2.2.0/src/itsdangerous/encoding.py#L50